Deferred: Add depth to UAV inputs in compute shader
We create an additional UAV, copy the depth buffer into it, and use it in the compute shader.

Note: No hit is equal to a 1 in the depth buffer. The other values were in the 0.99... range. So I transformed the values first before displaying them.

You can tell the difference more easily by zooming in and out.