FIX: Adjust Depth Buffer Mip Levels
Mip levels come in dimensions in powers of two. So 1x1, 2x2, 4x4 and so on.
The mip levels depend on the largest dimension of the texture and how many bits it requires. After resizing, this might change.

For instance, we initially have a height of 480 and a width of 640. The width is the larger number, and log2(640) is - if we round up - 10.
If we resize and now have a width of 1025, we need 11 mip levels, an additional one for the 2048 x 2048 level.
Similarly, if the height is now the largest dimension, we need to take that into
account instead of the width.